feat(front): display trial message usage in agent builder testing tab
Show trial message usage progress bar and info message in the agent
builder's Testing tab to inform users that testing consumes their
trial message credits.

Changes:
- Extract TrialMessageUsage component to shared location for reuse
- Add useIsTrialPlanWithMessageLimit hook to detect trial plans
- Display info message and progress bar at top of Testing tab
- Refactor NavigationSidebar to use shared TrialMessageUsage component

Resolves: dust-tt/tasks#5966
