+++ front/components/navigation/NavigationSidebar.tsx
@@ -1,5 +1,4 @@
 import {
-  Button,
   classNames,
   cn,
   CollapseButton,
@@ -17,11 +16,11 @@ import { useRouter } from "next/router";
 import React, {
   useCallback,
   useContext,
-  useEffect,
   useMemo,
   useState,
 } from "react";
 
+import { TrialMessageUsage } from "@app/components/app/TrialMessageUsage";
 import { useWelcomeTourGuide } from "@app/components/assistant/WelcomeTourGuideProvider";
 import type { SidebarNavigation } from "@app/components/navigation/config";
 import { getTopNavigationTabs } from "@app/components/navigation/config";
@@ -30,9 +29,7 @@ import { useNavigationLoading } from "@app/components/sparkle/NavigationLoadingC
 import { SidebarContext } from "@app/components/sparkle/SidebarContext";
 import { UserMenu } from "@app/components/UserMenu";
 import type { AppStatus } from "@app/lib/api/status";
-import { AGENT_MESSAGE_COMPLETED_EVENT } from "@app/lib/notifications/events";
 import { FREE_TRIAL_PHONE_PLAN_CODE } from "@app/lib/plans/plan_codes";
-import { useTrialMessageUsage } from "@app/lib/swr/trial_message_usage";
 import { useAppStatus } from "@app/lib/swr/useAppStatus";
 import { useFeatureFlags } from "@app/lib/swr/workspaces";
 import type {
@@ -201,7 +198,9 @@ export const NavigationSidebar = React.forwardRef<
       </div>
       <div className="flex grow flex-col">{children}</div>
       {subscription.plan.code === FREE_TRIAL_PHONE_PLAN_CODE && (
-        <TrialMessageUsage isAdmin={isAdmin(owner)} workspaceId={owner.sId} />
+        <div className="mx-3 mb-3">
+          <TrialMessageUsage isAdmin={isAdmin(owner)} workspaceId={owner.sId} />
+        </div>
       )}
       {user && (
         <div
@@ -331,92 +330,6 @@ function SubscriptionPastDueBanner() {
         </>
       }
     />
-  );
-}
-
-const MESSAGE_USAGE_CRITICAL_THRESHOLD = 0.9;
-
-interface TrialMessageUsageProps {
-  isAdmin: boolean;
-  workspaceId: string;
-}
-
-function TrialMessageUsage({ isAdmin, workspaceId }: TrialMessageUsageProps) {
-  const { messageUsage, mutateMessageUsage } = useTrialMessageUsage({
-    workspaceId,
-  });
-
-  useEffect(() => {
-    const handleAgentMessageCompleted = () => {
-      void mutateMessageUsage();
-    };
-    window.addEventListener(
-      AGENT_MESSAGE_COMPLETED_EVENT,
-      handleAgentMessageCompleted
-    );
-    return () => {
-      window.removeEventListener(
-        AGENT_MESSAGE_COMPLETED_EVENT,
-        handleAgentMessageCompleted
-      );
-    };
-  }, [mutateMessageUsage]);
-
-  if (!messageUsage || messageUsage.limit === -1) {
-    return null;
-  }
-
-  const { count, limit } = messageUsage;
-  const percentage = limit > 0 ? count / limit : 0;
-  const isCritical = percentage >= MESSAGE_USAGE_CRITICAL_THRESHOLD;
-  const isAtLimit = count >= limit;
-
-  return (
-    <div
-      className={cn(
-        "mx-3 mb-3 rounded-lg border p-3",
-        "border-border dark:border-border-night",
-        "bg-background dark:bg-background-night"
-      )}
-    >
-      <div className="mb-2 flex items-center justify-between text-sm">
-        <span className="font-semibold text-foreground dark:text-foreground-night">
-          Trial messages used
-        </span>
-        <span className="font-medium text-foreground dark:text-foreground-night">
-          <span className={cn(isCritical && "text-red-600 dark:text-red-400")}>
-            {count}
-          </span>{" "}
-          / {limit}
-        </span>
-      </div>
-      <div
-        className={cn(
-          "h-2 w-full overflow-hidden rounded-full",
-          "bg-gray-100 dark:bg-gray-100-night"
-        )}
-      >
-        <div
-          className={cn(
-            "h-full rounded-full transition-all",
-            isCritical
-              ? "bg-red-700 dark:bg-red-700-night"
-              : "bg-foreground dark:bg-foreground-night"
-          )}
-          style={{ width: `${Math.min(percentage * 100, 100)}%` }}
-        />
-      </div>
-      {isAtLimit && isAdmin && (
-        <div className="mt-3">
-          <Link
-            href={`/w/${workspaceId}/subscription`}
-            className="no-underline"
-          >
-            <Button label="Subscribe to Dust" variant="primary" />
-          </Link>
-        </div>
-      )}
-    </div>
   );
 }
